prepare to move to more modular tests
