+++ tests/test_modules/info.py
