+++ tests/test_modules/install.py
