+++ tests/test_modules/patch.py
