+++ tests/test_modules/remove.py
