+++ tests/test_modules/update.py
